relval_steps holds steps for first run2 worfklow; a lot of stuff factorized away into MatrixUtil
