+++ Configuration/PyReleaseValidation/python/relval_steps.py
@@ -1,118 +1,4 @@
-class Matrix(dict):
-    def __setitem__(self,key,value):
-        if key in self:
-            print "ERROR in Matrix"
-            print "overwritting",key,"not allowed"
-        else:
-            self.update({float(key):WF(float(key),value)})
-
-    def addOverride(self,key,override):
-        self[key].addOverride(override)
-            
-#the class to collect all possible steps
-class Steps(dict):
-    def __setitem__(self,key,value):
-        if key in self:
-            print "ERROR in Step"
-            print "overwritting",key,"not allowed"
-            import sys
-            sys.exit(-9)
-        else:
-            self.update({key:value})
-            # make the python file named <step>.py
-            #if not '--python' in value:                self[key].update({'--python':'%s.py'%(key,)})
-
-    def overwrite(self,keypair):
-        value=self[keypair[1]]
-        self.update({keypair[0]:value})
-        
-class WF(list):
-    def __init__(self,n,l):
-        self.extend(l)
-        self.num=n
-        #the actual steps of this WF
-        self.steps=[]
-        self.overrides={}
-    def addOverride(self,overrides):
-        self.overrides=overrides
-        
-    def interpret(self,stepsDict):
-        for s in self:
-            print 'steps',s,stepsDict[s]
-            steps.append(stepsDict[s])
-
-    
-InputInfoNDefault=2000000    
-class InputInfo(object):
-    def __init__(self,dataSet,label='',run=[],files=1000,events=InputInfoNDefault,split=10,location='CAF',ib_blacklist=None,ib_block=None) :
-        self.run = run
-        self.files = files
-        self.events = events
-        self.location = location
-        self.label = label
-        self.dataSet = dataSet
-        self.split = split
-        self.ib_blacklist = ib_blacklist
-        self.ib_block = ib_block
-        
-    def das(self, das_options):
-        if len(self.run) is not 0:
-            command = ";".join(["das_client.py %s --query '%s'" % (das_options, query) for query in self.queries()])
-            command = "({0})".format(command)
-        else:
-            command = "das_client.py %s --query '%s'" % (das_options, self.queries()[0])
-       
-        # Run filter on DAS output 
-        if self.ib_blacklist:
-            command += " | grep -E -v "
-            command += " ".join(["-e '{0}'".format(pattern) for pattern in self.ib_blacklist])
-        command += " | sort -u"
-        return command
-
-    def lumiRanges(self):
-        if len(self.run) != 0:
-            return "echo '{\n"+",".join(('"%d":[[1,268435455]]\n'%(x,) for x in self.run))+"}'"
-        return None
-
-    def queries(self):
-        query_by = "block" if self.ib_block else "dataset"
-        query_source = "{0}#{1}".format(self.dataSet, self.ib_block) if self.ib_block else self.dataSet
-        if len(self.run) is not 0:
-            return ["file {0}={1} run={2} site=T2_CH_CERN".format(query_by, query_source, query_run) for query_run in self.run]
-        else:
-            return ["file {0}={1} site=T2_CH_CERN".format(query_by, query_source)]
-
-    def __str__(self):
-        if self.ib_block:
-            return "input from: {0} with run {1}#{2}".format(self.dataSet, self.ib_block, self.run)
-        return "input from: {0} with run {1}".format(self.dataSet, self.run)
-    
-# merge dictionaries, with prioty on the [0] index
-def merge(dictlist,TELL=False):
-    import copy
-    last=len(dictlist)-1
-    if TELL: print last,dictlist
-    if last==0:
-        # ONLY ONE ITEM LEFT
-        return copy.copy(dictlist[0])
-    else:
-        reducedlist=dictlist[0:max(0,last-1)]
-        if TELL: print reducedlist
-        # make a copy of the last item
-        d=copy.copy(dictlist[last])
-        # update with the last but one item
-        d.update(dictlist[last-1])
-        # and recursively do the rest
-        reducedlist.append(d)
-        return merge(reducedlist,TELL)
-
-def remove(d,key,TELL=False):
-    import copy
-    e = copy.deepcopy(d)
-    if TELL: print "original dict, BEF: %s"%d
-    del e[key]
-    if TELL: print "copy-removed dict, AFT: %s"%e
-    return e
+from MatrixUtil import *
 
 # step1 gensim: for run1
 step1Defaults = {'--relval'      : None, # need to be explicitly set
@@ -134,7 +20,6 @@ step1Up2015Defaults = {'-s' : 'GEN,SIM',
                              }
 
 steps = Steps()
-#wmsplit = {}
 
 #### Production test section ####
 steps['ProdMinBias']=merge([{'cfg':'MinBias_8TeV_pythia8_TuneCUETP8M1_cff','--relval':'9000,300'},step1Defaults])
@@ -142,8 +27,6 @@ steps['ProdTTbar']=merge([{'cfg':'TTbar_8TeV_TuneCUETP8M1_cfi','--relval':'9000,
 steps['ProdQCD_Pt_3000_3500']=merge([{'cfg':'QCD_Pt_3000_3500_8TeV_TuneCUETP8M1_cfi','--relval':'9000,50'},step1Defaults])
 
 #### data ####
-#list of run to harvest for 2010A: 144086,144085,144084,144083,144011,139790,139789,139788,139787,138937,138934,138924,138923
-#list of run to harvest for 2010B: 149442,149291,149181,149011,148822,147929,147115,146644
 Run2010ASk=[138937,138934,138924,138923,139790,139789,139788,139787,144086,144085,144084,144083,144011]
 Run2010BSk=[146644,147115,147929,148822,149011,149181,149182,149291,149294,149442]
 steps['MinimumBias2010A']={'INPUT':InputInfo(dataSet='/MinimumBias/Run2010A-valskim-v6/RAW-RECO',label='run2010A',location='STD',run=Run2010ASk)}
@@ -246,16 +129,9 @@ steps['WElSkim2012D']={'INPUT':InputInfo(dataSet='/SingleElectron/Run2012D-WElec
 steps['ZElSkim2012D']={'INPUT':InputInfo(dataSet='/DoubleElectron/Run2012D-ZElectron-PromptSkim-v1/RAW-RECO',label='zEl2012D',location='STD',run=Run2012Dsk)}
 
 #### run2 ####
-Run2015B=[251642] #  251561 251638 251642
-steps['RunHLTPhy2015B']={'INPUT':InputInfo(dataSet='/HLTPhysics/Run2015B-v1/RAW',label='hltPhy2015B',run=Run2015B, events=100000,location='STD')}
-
-#### Standard release validation samples ####
-
-stCond={'--conditions':'auto:run1_mc'}
-def Kby(N,s):
-    return {'--relval':'%s000,%s'%(N,s)}
-def Mby(N,s):
-    return {'--relval':'%s000000,%s'%(N,s)}
+# Run2015B=[251642] #  251561 251638 251642
+Run2015B=selectedLS([251251])
+steps['RunHLTPhy2015B']={'INPUT':InputInfo(dataSet='/HLTPhysics/Run2015B-v1/RAW',label='hltPhy2015B',events=100000,location='STD', ls=Run2015B)}
 
 def gen(fragment,howMuch):
     global step1Defaults
@@ -592,25 +468,6 @@ steps['PhotonJets_Pt_10_13_HI']=merge([hiDefaults,steps['PhotonJets_Pt_10_13']])
 steps['ZMM_13_HI']=merge([hiDefaults,steps['ZMM_13']])
 steps['ZEEMM_13_HI']=merge([hiDefaults,steps['ZEEMM_13']])
 
-
-def changeRefRelease(steps,listOfPairs):
-    for s in steps:
-        if ('INPUT' in steps[s]):
-            oldD=steps[s]['INPUT'].dataSet
-            for (ref,newRef) in listOfPairs:
-                if  ref in oldD:
-                    steps[s]['INPUT'].dataSet=oldD.replace(ref,newRef)
-        if '--pileup_input' in steps[s]:
-            for (ref,newRef) in listOfPairs:
-                if ref in steps[s]['--pileup_input']:
-                    steps[s]['--pileup_input']=steps[s]['--pileup_input'].replace(ref,newRef)
-        
-def addForAll(steps,d):
-    for s in steps:
-        steps[s].update(d)
-
-
-
 #### fastsim section ####
 ##no forseen to do things in two steps GEN-SIM then FASTIM->end: maybe later
 step1FastDefaults =merge([{'-s':'GEN,SIM,RECOBEFMIX,DIGI:pdigi_valid,L1,L1Reco,RECO,EI,HLT:@fake,VALIDATION',
@@ -743,21 +600,6 @@ step1LHEDefaults=merge([{'-s':'LHE',
                          '--conditions':'auto:run2_mc_FULL'                         
                          },
                         step1Defaults])
-
-
-def genvalid(fragment,d,suffix='all',fi='',dataSet=''):
-    import copy
-    c=copy.copy(d)
-    if suffix:
-        c['-s']=c['-s'].replace('genvalid','genvalid_'+suffix)
-    if fi:
-        c['--filein']='lhe:%d'%(fi,)
-    if dataSet:
-        c['--filein']='das:%s'%(dataSet,)
-    c['cfg']=fragment
-    return c
-
-
 
 steps['DYToll01234Jets_5f_LO_MLM_Madgraph_LHE_13TeV']=genvalid('Configuration/Generator/python/DYToll01234Jets_5f_LO_MLM_Madgraph_LHE_13TeV_cff.py',step1LHEDefaults)
 steps['TTbar012Jets_5f_NLO_FXFX_Madgraph_LHE_13TeV']=genvalid('Configuration/Generator/python/TTbar012Jets_5f_NLO_FXFX_Madgraph_LHE_13TeV_cff.py',step1LHEDefaults)
@@ -947,7 +789,6 @@ steps['DIGIPRMXUP15_PROD_PU50']=merge([premixProd50ns,digiPremixUp2015Defaults50
 dataReco={ '--runUnscheduled':'',
           '--conditions':'auto:run1_data',
           '-s':'RAW2DIGI,L1Reco,RECO,EI,PAT,ALCA:SiStripCalZeroBias+SiStripCalMinBias+TkAlMinBias,DQM:@standardDQM+@miniAODDQM',
-          #'-s':'RAW2DIGI,L1Reco,RECO,EI,ALCA:SiStripCalZeroBias+SiStripCalMinBias+TkAlMinBias,DQM:@standardDQM+@miniAODDQM',  # re-vert PAT away if there are issues
           '--datatier':'RECO,MINIAOD,DQMIO',
           '--eventcontent':'RECO,MINIAOD,DQM',
           '--data':'',
@@ -974,15 +815,12 @@ steps['RECODAlCaCalo']=merge([{'--scenario':'pp',},dataRecoAlCaCalo])
 
 hltKey='frozen50ns'
 menuR2 = autoHLT[hltKey]
-
 # no GT customization for HLT frozen50ns
 steps['HLTDR2']=merge( [ {'-s':'L1REPACK,HLT:@%s'%hltKey,},{'--conditions':'auto:run2_hlt',},steps['HLTD'] ] )
 
-steps['RECODR2']=merge([{'--scenario':'pp','--conditions':'auto:run2_data',},dataReco])
-steps['RECODR2AlCaEle']=merge([{'--scenario':'pp','--conditions':'auto:run2_data',},dataRecoAlCaCalo])
 # custom function to be put back once the CSC tracked/untracked will have been fixed.. :-)
-#steps['RECODR2']=merge([{'--scenario':'pp','--conditions':'auto:run2_data','--customise':'Configuration/DataProcessing/RecoTLR.customisePromptRun2',},dataReco])
-#steps['RECODR2AlCaEle']=merge([{'--scenario':'pp','--conditions':'auto:run2_data','--customise':'Configuration/DataProcessing/RecoTLR.customisePromptRun2',},dataRecoAlCaCalo])
+steps['RECODR2']=merge([{'--scenario':'pp','--conditions':'auto:run2_data','--customise':'Configuration/DataProcessing/RecoTLR.customisePromptRun2',},dataReco])
+steps['RECODR2AlCaEle']=merge([{'--scenario':'pp','--conditions':'auto:run2_data','--customise':'Configuration/DataProcessing/RecoTLR.customisePromptRun2',},dataRecoAlCaCalo])
 
 
 steps['RECODSplit']=steps['RECOD'] # finer job splitting  
@@ -1244,7 +1082,7 @@ steps['HARVESTD']={'-s':'HARVESTING:@standardDQM+@miniAODDQM',
                    '--scenario':'pp'}
 
 steps['HARVESTDreHLT'] = merge([ {'--conditions':'auto:run1_data_%s'%menu}, steps['HARVESTD'] ])
-steps['HARVESTDR2reHLT'] = merge([ {'--conditions':'auto:run2_data'}, steps['HARVESTD'] ])
+steps['HARVESTDR2reHLT'] = merge([ {'--conditions':'auto:run2_data','--customise':'Configuration/DataProcessing/RecoTLR.customisePromptRun2',}, steps['HARVESTD'] ])
 
 steps['HARVESTDDQM']=merge([{'-s':'HARVESTING:@common+@muon+@hcal+@jetmet+@ecal'},steps['HARVESTD']])
 
@@ -1582,4 +1420,3 @@ for step in upgradeSteps:
             k=step+'_'+key
             if step in upgradeStepDict and key in upgradeStepDict[step]: 
                 steps[k]=merge([upgradeStepDict[step][key]])
-
